src/requirements.txt
click==8.0.3
Flask==2.0.2
Flask-WTF==0.15.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mock==4.0.3
mysql-connector-python==8.0.29
python-dotenv==0.19.1 
pylint==2.13.4
pytest==7.1.1
coverage==6.2
pytest-cov==3.0.0
pytest-asyncio==0.18.3
prometheus-flask-exporter==0.20.1
requests==2.26.0
Werkzeug==2.0.2
